Change avatar_version to long of current time; Fix the bug changing avatar causing a crash; Implement follow in following lists.

+++ LearnLeague/app/src/main/java/edu/cuhk/csci3310/learnleague/ApiClient.java
@@ -171,7 +171,7 @@ public class ApiClient {
                                 jsonResponse.getString("hashedpassword"),
                                 jsonResponse.getString("username"),
                                 jsonResponse.getString("email"),
-                                jsonResponse.getInt("avatarversion")
+                                jsonResponse.getLong("avatarversion")
                         );
                     } catch (JSONException e) {
                         throw new RuntimeException(e);
